Fix resolution of deploy-service url
When there are more than one shared domains, presented in the landscape,
the logic of finding the correct one, was assuming that the first one is
the correct. However, this is not true all the times as the first one
could not be the correct one. This commit introduces a mechanism for
pinging the domains till the right one responds with 200 OK status code.

BUG: LMCROSSITXSADEPLOY-833